implemented user-recipe association

+++ app/controllers/recipes_controller.rb
@@ -8,6 +8,7 @@ class RecipesController < ApplicationController
   end
 
   def new
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
     @recipe = Recipe.new
     1.times { @recipe.ingredients.build }
     1.times { @recipe.instructions.build }
@@ -15,9 +16,9 @@ class RecipesController < ApplicationController
 
   def create
     @recipe = Recipe.new(recipe_params)
+    @recipe.user = current_user
 
     if @recipe.save
-
       redirect_to @recipe
     else
       render :new
@@ -25,7 +26,13 @@ class RecipesController < ApplicationController
   end
 
   def edit
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
+
     @recipe = Recipe.find(params[:id])
+
+    if logged_in?
+      session_notice(:danger, 'Wrong user') unless equal_with_current_user?(@recipe.user)
+    end
   end
 
   def update
@@ -39,10 +46,15 @@ class RecipesController < ApplicationController
   end
 
   def destroy
+    session_notice(:danger, 'You must be logged in!') unless logged_in?
     recipe = Recipe.find(params[:id])
-    recipe.destroy
 
-    redirect_to root_path
+    if equal_with_current_user?(recipe.user)
+      recipe.destroy
+      redirect_to recipes_path
+    else
+      session_notice(:danger, 'Wrong User')
+    end
   end
 
   private
